Add /hidealtstext command (#3750)
When a user is spamming, often times they are doing so under several accounts.  When this happens, a staff member has to manually hide each alts messages in the room, which takes up one line of the chat for each alt.  This command makes these cases easier and less-spammy for the chat.