Make timer track in seconds rather than ticks
In the past, I've gotten burnt out trying to implement a realtime
timer, so this is going to take a different and more granular approach.

The first step is this: Just making the variables track in seconds.
Everything still _happens_ in ticks, it's just the variables store
seconds rather than ticks.