Give Hosts the forcewin permission
This allows groupchat tournament hosts, as well as `/importinputlog`
users, to use the `/forcewin` and `/editbattle` commands.

This started as just wanting to give `/editbattle reseed` to users of
`/importinputlog`, but in hindsight, all these permissions could be
useful and allow for more creative tournaments, so I'm opening it up
fully, and we'll see how everyone does with them.