BDSP: Remove formats with under 500 games
There's not really a point in keeping these metas if they don't get traction on the ladder, don't get attention in the room/forum, and no longer track usage for tiering; plus, modified rulesets can be used in challenges anyway. These formats would be removed with the addition of Scarlet/Violet anyway.