Fix crash in Rulesets
Crash introduced in #8195 and has been plaguing us with thousands of
crashes per month for the past two years, but b319c1f6b0aeecf finally
let us track it down.

Special thanks to everyone who spent so much time and effort trying
to identify the source of this crash over the past few years. It's
over. It's finally over.