Exclude unobtainable HM moves from movesearch (#10598)
* Exclude unobtainable HM moves from movesearch

https://www.smogon.com/forums/threads/impossible-hm-moves-in-movesearch.3735989/

* Update dex-species.ts

* Update dex-species.ts

* Update dex-species.ts

* Properly deal with CAP

After doing some digging, I found that 08bca77cf48773cd284dcf72bb2a96094b5898dc is when the isNonstandard check was added to allow CAP Pokemon to know HM moves. This was untouched for over ten years and was inadvertently allowing dexited Pokemon to have HM moves in NatDex that should be illegal in most other formats.

I don't believe this was ever intentional based on knowing the general mechanics of NatDex, nor does it have any impact on the metagame.