Refactor spotlights
Fix /removedaily, adding a new "all" option to work like it's
documented to work, and fix the documentation.

Also remove support for legacy image format. A lot of functions were
async that didn't need to be.

Also remove an entirely unused sorting pass mistakenly introduced in
8f4533ff2171900ef57a0a9b0d
(The queue was sorted, then immediately ignored to iterate it by index,
which is for the best, because showing the queue in any order other
than queue order is not useful.)