SV OU: Add Evasion Abilities Clause (#11175)
Replaces the manual bans of Snow Cloak and Sand Veil with Evasion Abilities Clause. There is no actual impact on legality, this is a consistency change. This change is approved by star and finch:

starmaster — 10:32 AM
we did literally create it for ou
so idk why that is
it should just have the clause
UT ⛵ — 10:35 AM
@Finch do you mind if I submit a pull request to remove Sand Veil + Snow Cloak bans from OU and replace it with Evasion Abilities Clause? This should have absolutely no change on the tier and is a backend consistency thing
Finch — 10:36 AM
@Ruft good with you? Checks out to me